refactor(scene): split UI scene IR responsibilities
